% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{readability_chr}
\alias{readability_chr}
\title{Get readability stats for selected text (excluding code chunks)}
\usage{
readability_chr(text)
}
\arguments{
\item{text}{a character string of text, length of one}
}
\description{
Use this function with a character string as input
}
