% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{text_stats}
\alias{text_stats}
\title{Get text stats for selected text (excluding code chunks and inline code)}
\usage{
text_stats()
}
\description{
Call this addin to get a word count and some other stats about the text
}

