% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{text_stats_chr}
\alias{text_stats_chr}
\title{Get text stats for selected text (excluding code chunks and inline code)}
\usage{
text_stats_chr(text)
}
\description{
Use this function with a character string as input
}
